requirements.txt
fastapi
uvicorn
python-dotenv
pandas
openpyxl
chromadb
openai
sentence-transformers
pytest
